+++ frontend/src/service/@shared/hooks/queries/auth/index.ts
@@ -1,0 +1,3 @@
+export * from './useGet';
+export * from './useCreate';
+export * from './useDelete';
